♻️ refactor(types): use u8 for RuntimeSignal::Exit and ProjectRunOutcome::Exited
Exit codes are always 0..=255 (POSIX). Using u8 in the type signature
lets the type system enforce the constraint instead of a runtime check.

- RuntimeSignal::Exit(i64) → Exit(u8)
- ProjectRunOutcome::Exited(i64) → Exited(u8)
- require_exit_code now returns u8 via u8::try_from
- project_exit_code simplified: no longer needs Result, just ExitCode::from(u8)
- Call site in main.rs simplified accordingly

+++ crates/sporec-codegen/src/lib.rs
@@ -33,7 +33,7 @@ pub enum SpecKind {
 #[derive(Debug, Clone, PartialEq)]
 pub enum ProjectRunOutcome {
     Completed(Value),
-    Exited(i64),
+    Exited(u8),
 }
 
 /// Execute a Spore module by calling its current default startup function
